JBRULES-1745 WS-Human-Task -Changed OrganizationalEntity id from long to String
git-svn-id: http://anonsvn.jboss.org/repos/labs/labs/jbossrules/trunk@22786 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-process/drools-process-task/src/test/java/org/drools/task/service/MockEventMessagingTest.java
@@ -48,7 +48,7 @@ public class MockEventMessagingTest extends BaseTest {
         
         assertEquals( 1, transport.list.size() );
         assertEquals( taskId, ((TaskClaimedEvent) ((Payload) transport.list.get(0)).get()).getTaskId() );
-        assertEquals( users.get( "darth" ).getId(), (Long) ((TaskClaimedEvent) ((Payload) transport.list.get(0)).get()).getUserId() );
+        assertEquals( users.get( "darth" ).getId(), ((TaskClaimedEvent) ((Payload) transport.list.get(0)).get()).getUserId() );
         
     }
